+++ YukiFrameWork/Framework/Extension/Mono/SceneTool.cs
@@ -13,12 +13,12 @@ using System.Collections;
 using YukiFrameWork.Extension;
 using XFABManager;
 namespace YukiFrameWork
-{
+{   
 	public static class SceneTool
 	{	
 		public static event Action<float> LoadingScene = null;
         public static event Action LoadSceneSucceed = null;
-
+       
         public static readonly XFABManagerSceneTool XFABManager = new XFABManagerSceneTool();
 
 		public static void LoadScene(string sceneName,LoadSceneMode mode = LoadSceneMode.Single)
@@ -154,9 +154,14 @@ namespace YukiFrameWork
             if (!isInited) return;
             AssetBundleManager.LoadScene(projectName, sceneName, mode);
         }
-        [DisableEnumeratorWarning]
+
         public IEnumerator LoadSceneAsync(string sceneName, Action<float> loadingCallBack = null, LoadSceneMode mode = LoadSceneMode.Single)
             => isInited ? SceneTool.LoadSceneAsync(AssetBundleManager.LoadSceneAsynchrony(projectName, sceneName, mode),loadingCallBack) : throw new Exception("没有完成对SceneTool.XFABManager的初始化，请调用一次Init方法");
+
+        public LoadSceneRequest LoadSceneAsync(string sceneName, LoadSceneMode loadSceneMode = LoadSceneMode.Single)
+        {
+            return isInited ? AssetBundleManager.LoadSceneAsynchrony(projectName, sceneName, loadSceneMode) : throw new Exception("没有完成对SceneTool.XFABManager的初始化，请调用一次Init方法");
+        }
 
         /// <summary>
         /// 场景安全加载，如果场景就是当前需要切换的场景，那么会直接执行OnFinish回调，不会执行场景的Load逻辑
@@ -174,6 +179,25 @@ namespace YukiFrameWork
             }           
             await MonoHelper.Start(LoadSceneAsync(sceneName, loadingCallBack,mode));
             onFinish?.Invoke();
+        }
+        /// <summary>
+        /// 场景安全加载，如果场景就是当前需要切换的场景，那么会直接执行OnFinish回调，不会执行场景的Load逻辑
+        /// </summary>
+        /// <param name="sceneName"></param>
+        /// <param name="loadingCallBack"></param>
+        /// <param name="onFinish"></param>
+        /// <param name="mode"></param>
+        public async void LoadSceneAsyncSafe(string sceneName, Action<float> loadingCallBack = null, Action<Scene> onFinish = null, LoadSceneMode mode = LoadSceneMode.Single)
+        {
+            Scene scene = SceneManager.GetSceneByName(sceneName);
+            if (scene.IsValid())
+            {
+                onFinish?.Invoke(scene);
+                return;
+            }
+            var request = LoadSceneAsync(sceneName, mode);
+            await request;
+            onFinish?.Invoke(request.Scene);
         }
 
         [DisableEnumeratorWarning]